// base/interfaces.ts
export type IShadowRootMode = 'closed' | 'open';

export type ISupportedType = 'application/xhtml+xml' | 'application/xml' | 'image/svg+xml' | 'text/html' | 'text/xml';

export type ITextTrackKind = 'captions' | 'chapters' | 'descriptions' | 'metadata' | 'subtitles';

export type ITextTrackMode = 'disabled' | 'hidden' | 'showing';

export type IVisibilityState = 'hidden' | 'prerender' | 'visible';

export type IEventListenerOrEventListenerObject = IEventListener | IEventListenerObject;
// ...
export interface IAddEventListenerOptions extends IEventListenerOptions {
  once?: boolean;
  passive?: boolean;
}
// ...
export interface IEventListenerOptions {
  capture?: boolean;
}

export interface IFilePropertyBag extends IBlobPropertyBag {
  lastModified?: number;
}

export interface IGetRootNodeOptions {
  composed?: boolean;
}

export interface IShadowRootInit {
  mode: IShadowRootMode;
}

export interface IStaticRangeInit {
  endContainer: INode;
  endOffset: number;
  startContainer: INode;
  startOffset: number;
}

export type IBlobCallback = (blob: IBlob | null) => void;

export type IEventListener = (evt: IEvent) => void;
// ...
export interface IAttr extends INode {
  readonly localName: string;
  readonly name: string;
  readonly namespaceURI: string | null;
  readonly ownerElement: IElement | null;
  readonly prefix: string | null;
  readonly specified: boolean;
  value: string;
}
// ...
export interface IChildNode {
  after(...nodes: (INode | string)[]): void;
  before(...nodes: (INode | string)[]): void;
  remove(): void;
  replaceWith(...nodes: (INode | string)[]): void;
}
// ...
export interface IDOMRect extends IDOMRectReadOnly {
  height: number;
  width: number;
  x: number;
  y: number;
}

export interface IDOMRectList {
  readonly length: number;
  item(index: number): IDOMRect | null;
  [index: number]: IDOMRect;
}

export interface IDOMRectReadOnly {
  readonly bottom: number;
  readonly height: number;
  readonly left: number;
  readonly right: number;
  readonly top: number;
  readonly width: number;
  readonly x: number;
  readonly y: number;
  toJSON(): any;
}

export interface IDOMStringMap {
  [name: string]: string;
}

export interface IDOMTokenList {
  readonly length: number;
  value: string;
  add(...tokens: string[]): void;
  contains(token: string): boolean;
  item(index: number): string | null;
  remove(...tokens: string[]): void;
  replace(token: string, newToken: string): boolean;
  supports(token: string): boolean;
  toggle(token: string, force?: boolean): boolean;
  toString(): string;
  forEach(callbackfn: (value: string, key: number, parent: IDOMTokenList) => void, thisArg?: any): void;
  [index: number]: string;
}
// ...
export interface IDocumentFragment extends INode, INonElementParentNode, IParentNode {
}

export interface IDocumentOrShadowRoot {
  readonly activeElement: IElement | null;
}

export interface IDocumentType extends INode, IChildNode {
  readonly name: string;
  readonly publicId: string;
  readonly systemId: string;
}

export interface IElement extends INode, IChildNode, INonDocumentTypeChildNode, IParentNode, ISlotable {
  readonly attributes: INamedNodeMap;
  readonly classList: IDOMTokenList;
  className: string;
  readonly clientHeight: number;
  readonly clientLeft: number;
  readonly clientTop: number;
  readonly clientWidth: number;
  id: string;
  innerHTML: string;
  readonly localName: string;
  readonly namespaceURI: string | null;
  outerHTML: string;
  readonly prefix: string | null;
  readonly scrollHeight: number;
  scrollLeft: number;
  scrollTop: number;
  readonly scrollWidth: number;
  readonly shadowRoot: IShadowRoot | null;
  slot: string;
  readonly tagName: string;
  attachShadow(init: IShadowRootInit): IShadowRoot;
  closest(selectors: string): IElement | null;
  getAttribute(qualifiedName: string): string | null;
  getAttributeNS(namespace: string | null, localName: string): string | null;
  getAttributeNames(): Iterable<string>;
  getAttributeNode(qualifiedName: string): IAttr | null;
  getAttributeNodeNS(namespace: string | null, localName: string): IAttr | null;
  getBoundingClientRect(): IDOMRect;
  getClientRects(): IDOMRectList;
  getElementsByClassName(classNames: string): IHTMLCollection;
  getElementsByTagName(qualifiedName: string): IHTMLCollection;
  getElementsByTagNameNS(namespace: string | null, localName: string): IHTMLCollection;
  hasAttribute(qualifiedName: string): boolean;
  hasAttributeNS(namespace: string | null, localName: string): boolean;
  hasAttributes(): boolean;
  insertAdjacentElement(where: string, element: IElement): IElement | null;
  insertAdjacentHTML(position: string, text: string): void;
  insertAdjacentText(where: string, data: string): void;
  matches(selectors: string): boolean;
  removeAttribute(qualifiedName: string): void;
  removeAttributeNS(namespace: string | null, localName: string): void;
  removeAttributeNode(attr: IAttr): IAttr;
  setAttribute(qualifiedName: string, value: string): void;
  setAttributeNS(namespace: string | null, qualifiedName: string, value: string): void;
  setAttributeNode(attr: IAttr): IAttr | null;
  setAttributeNodeNS(attr: IAttr): IAttr | null;
  toggleAttribute(qualifiedName: string, force?: boolean): boolean;
  webkitMatchesSelector(selectors: string): boolean;
}

export interface IElementCSSInlineStyle {
  readonly style: ICSSStyleDeclaration;
}

export interface IElementContentEditable {
}
// ...
export interface IEventListenerObject {
  handleEvent(evt: IEvent): void;
}

export interface IEventTarget {
  dispatchEvent(event: IEvent): boolean;
}

export interface IFile extends IBlob {
  readonly lastModified: number;
  readonly name: string;
}

export interface IFileList {
  readonly length: number;
  item(index: number): IFile | null;
  [index: number]: IFile;
}

export interface IGlobalEventHandlersEventMap {
  abort: IEvent;
  auxclick: IEvent;
  blur: IEvent;
  cancel: IEvent;
  canplay: IEvent;
  canplaythrough: IEvent;
  change: IEvent;
  click: IEvent;
  close: IEvent;
  contextmenu: IEvent;
  cuechange: IEvent;
  dblclick: IEvent;
  drag: IEvent;
  dragend: IEvent;
  dragenter: IEvent;
  dragexit: IEvent;
  dragleave: IEvent;
  dragover: IEvent;
  dragstart: IEvent;
  drop: IEvent;
  durationchange: IEvent;
  emptied: IEvent;
  ended: IEvent;
  error: IEvent;
  focus: IEvent;
  input: IEvent;
  invalid: IEvent;
  keydown: IEvent;
  keypress: IEvent;
  keyup: IEvent;
  load: IEvent;
  loadeddata: IEvent;
  loadedmetadata: IEvent;
  loadend: IEvent;
  loadstart: IEvent;
  mousedown: IEvent;
  mouseenter: IEvent;
  mouseleave: IEvent;
  mousemove: IEvent;
  mouseout: IEvent;
  mouseover: IEvent;
  mouseup: IEvent;
  pause: IEvent;
  play: IEvent;
  playing: IEvent;
  progress: IEvent;
  ratechange: IEvent;
  reset: IEvent;
  resize: IEvent;
  scroll: IEvent;
  securitypolicyviolation: IEvent;
  seeked: IEvent;
  seeking: IEvent;
  select: IEvent;
  stalled: IEvent;
  submit: IEvent;
  suspend: IEvent;
  timeupdate: IEvent;
  toggle: IEvent;
  volumechange: IEvent;
  waiting: IEvent;
  wheel: IEvent;
}
// ...
export interface IHTMLCollection<T extends IElement = IElement> extends IHTMLCollectionBase {
  namedItem(name: string): T | null;
}

export interface IHTMLCollectionBase {
  readonly length: number;
  item(index: number): IElement | null;
  [index: number]: IElement;
}
// ...
export interface IHTMLOrSVGElement {
  readonly dataset: IDOMStringMap;
  nonce: string;
  tabIndex: number;
  blur(): void;
  focus(): void;
}
// ...
export interface INamedNodeMap {
  readonly length: number;
  getNamedItem(qualifiedName: string): IAttr | null;
  getNamedItemNS(namespace: string | null, localName: string): IAttr | null;
  item(index: number): IAttr | null;
  removeNamedItem(qualifiedName: string): IAttr;
  removeNamedItemNS(namespace: string | null, localName: string): IAttr;
  setNamedItem(attr: IAttr): IAttr | null;
  setNamedItemNS(attr: IAttr): IAttr | null;
  [index: number]: IAttr;
}

export interface INode extends IEventTarget {
  readonly ATTRIBUTE_NODE: number;
  readonly CDATA_SECTION_NODE: number;
  readonly COMMENT_NODE: number;
  readonly DOCUMENT_FRAGMENT_NODE: number;
  readonly DOCUMENT_NODE: number;
  readonly DOCUMENT_POSITION_CONTAINED_BY: number;
  readonly DOCUMENT_POSITION_CONTAINS: number;
  readonly DOCUMENT_POSITION_DISCONNECTED: number;
  readonly DOCUMENT_POSITION_FOLLOWING: number;
  readonly DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC: number;
  readonly DOCUMENT_POSITION_PRECEDING: number;
  readonly DOCUMENT_TYPE_NODE: number;
  readonly ELEMENT_NODE: number;
  readonly ENTITY_NODE: number;
  readonly ENTITY_REFERENCE_NODE: number;
  readonly NOTATION_NODE: number;
  readonly PROCESSING_INSTRUCTION_NODE: number;
  readonly TEXT_NODE: number;
  readonly baseURI: string;
  readonly childNodes: INodeList;
  readonly firstChild: INode | null;
  readonly isConnected: boolean;
  readonly lastChild: INode | null;
  readonly nextSibling: INode | null;
  readonly nodeName: string;
  readonly nodeType: number;
  nodeValue: string | null;
  readonly ownerDocument: IDocument | null;
  readonly parentElement: IElement | null;
  readonly parentNode: INode | null;
  readonly previousSibling: INode | null;
  textContent: string | null;
  appendChild(node: INode): INode;
  cloneNode(deep?: boolean): INode;
  compareDocumentPosition(other: INode): number;
  contains(other: INode | null): boolean;
  getRootNode(options?: IGetRootNodeOptions): INode;
  hasChildNodes(): boolean;
  insertBefore(node: INode, child: INode | null): INode;
  isDefaultNamespace(namespace: string | null): boolean;
  isEqualNode(otherNode: INode | null): boolean;
  isSameNode(otherNode: INode | null): boolean;
  lookupNamespaceURI(prefix: string | null): string | null;
  lookupPrefix(namespace: string | null): string | null;
  normalize(): void;
  removeChild(child: INode): INode;
  replaceChild(node: INode, child: INode): INode;
}

export interface INodeIterator {
  readonly filter: INodeFilter | null;
  readonly pointerBeforeReferenceNode: boolean;
  readonly referenceNode: INode;
  readonly root: INode;
  readonly whatToShow: number;
  detach(): void;
  nextNode(): INode | null;
  previousNode(): INode | null;
}

export interface INodeList<T extends INode = INode> {
  readonly length: number;
  item(index: number): T | null;
  forEach(callbackfn: (value: INode, key: number, parent: INodeList<T>) => void, thisArg?: any): void;
  [index: number]: T;
}

export interface INonDocumentTypeChildNode {
  readonly nextElementSibling: IElement | null;
  readonly previousElementSibling: IElement | null;
}

export interface INonElementParentNode {
  getElementById(elementId: string): IElement | null;
}

export interface IParentNode {
  readonly childElementCount: number;
  readonly children: IHTMLCollection;
  readonly firstElementChild: IElement | null;
  readonly lastElementChild: IElement | null;
  append(...nodes: (INode | string)[]): void;
  prepend(...nodes: (INode | string)[]): void;
  querySelector(selectors: string): IElement | null;
  querySelectorAll(selectors: string): INodeList;
}
// ...
export interface IShadowRoot extends IDocumentFragment, IDocumentOrShadowRoot {
  readonly host: IElement;
  innerHTML: string;
  readonly mode: IShadowRootMode;
}

export interface ISlotable {
  readonly assignedSlot: IHTMLSlotElement | null;
}
// ...
export interface IHTMLElement extends IElement, IElementCSSInlineStyle, IElementContentEditable, IGlobalEventHandlers, IHTMLOrSVGElement {
  accessKey: string;
  dir: string;
  draggable: boolean;
  hidden: boolean;
  lang: string;
  readonly offsetHeight: number;
  readonly offsetLeft: number;
  readonly offsetParent: IElement | null;
  readonly offsetTop: number;
  readonly offsetWidth: number;
  title: string;
  translate: boolean;
  click(): void;
  addEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IAddEventListenerOptions): void;
  addEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IAddEventListenerOptions): void;
  removeEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IEventListenerOptions): void;
  removeEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IEventListenerOptions): void;
}
// ...
export interface IHTMLSlotElement extends IHTMLElement {
  name: string;
  assignedElements(options?: IAssignedNodesOptions): Iterable<IElement>;
  assignedNodes(options?: IAssignedNodesOptions): Iterable<INode>;
  addEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLSlotElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IAddEventListenerOptions): void;
  addEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IAddEventListenerOptions): void;
  removeEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLSlotElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IEventListenerOptions): void;
  removeEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IEventListenerOptions): void;
}